Do the steps in the exact match

Do the steps in the exact match

do_exactmatch(
  db_folder,
  data_intid,
  select_before = "2021-01-01",
  select_crops = NULL,
  match_type = "raw",
  rename_df = FALSE
)

Arguments

db_folder

A string path to the database controlled vocabulary folder

data_intid

A data.frame created by the function create_intid

select_before

A string in the format of Ymd. The function returns cultivars that were added to the datebase before this specified date.

select_crops

A regular expression of crop types separated by |

match_type

A string, either "raw" or "db, " which denotes the type of matching procedure to use. If matching raw variety names, use "raw." If matching database names, use "db." The difference is that "raw" groups by the var_id and is only looking for one match per var_id. The method "db" groups by the intid and is looking to match every entry provided.

rename_df

A logical denoting whether to check the rename (misspelling) file Otherwise, the matches are pulled from the database list

Currently, all the programs renames are kept in one file, but may want to consider putting the renames into different files that are collected together See get_cultivar_rename()

data_suffix

A character suffix to be appended to the columns related to the original data

See also